refactor: Fix CLI and migrate commands/report/execute.py to current_app.config
- Remove superset/cli/lib.py to avoid direct config imports
- Move normalize_token function directly into cli/main.py
- Update cli/reset.py to check feature flags using current_app.config
- Migrate commands/report/execute.py from app.config to current_app.config with conf alias
- This eliminates problematic direct imports from superset.config in CLI modules
