+++ superset/commands/report/execute.py
@@ -22,7 +22,7 @@ from uuid import UUID
 import pandas as pd
 from celery.exceptions import SoftTimeLimitExceeded
 
-from superset import app, db, security_manager
+from superset import db, security_manager
 from superset.commands.base import BaseCommand
 from superset.commands.dashboard.permalink.create import CreateDashboardPermalinkCommand
 from superset.commands.exceptions import CommandException, UpdateFailedError
@@ -310,18 +310,21 @@ class BaseReportState:
         Get chart or dashboard screenshots
         :raises: ReportScheduleScreenshotFailedError
         """
+        from flask import current_app
+
+        conf = current_app.config
         _, username = get_executor(
-            executors=app.config["ALERT_REPORTS_EXECUTORS"],
+            executors=conf["ALERT_REPORTS_EXECUTORS"],
             model=self._report_schedule,
         )
         user = security_manager.find_user(username)
 
-        max_width = app.config["ALERT_REPORTS_MAX_CUSTOM_SCREENSHOT_WIDTH"]
+        max_width = conf["ALERT_REPORTS_MAX_CUSTOM_SCREENSHOT_WIDTH"]
 
         if self._report_schedule.chart:
             url = self._get_url()
 
-            window_width, window_height = app.config["WEBDRIVER_WINDOW"]["slice"]
+            window_width, window_height = conf["WEBDRIVER_WINDOW"]["slice"]
             width = min(max_width, self._report_schedule.custom_width or window_width)
             height = self._report_schedule.custom_height or window_height
             window_size = (width, height)
@@ -331,13 +334,13 @@ class BaseReportState:
                     url,
                     self._report_schedule.chart.digest,
                     window_size=window_size,
-                    thumb_size=app.config["WEBDRIVER_WINDOW"]["slice"],
+                    thumb_size=conf["WEBDRIVER_WINDOW"]["slice"],
                 )
             ]
         else:
             urls = self.get_dashboard_urls()
 
-            window_width, window_height = app.config["WEBDRIVER_WINDOW"]["dashboard"]
+            window_width, window_height = conf["WEBDRIVER_WINDOW"]["dashboard"]
             width = min(max_width, self._report_schedule.custom_width or window_width)
             height = self._report_schedule.custom_height or window_height
             window_size = (width, height)
@@ -347,7 +350,7 @@ class BaseReportState:
                     url,
                     self._report_schedule.dashboard.digest,
                     window_size=window_size,
-                    thumb_size=app.config["WEBDRIVER_WINDOW"]["dashboard"],
+                    thumb_size=conf["WEBDRIVER_WINDOW"]["dashboard"],
                 )
                 for url in urls
             ]
@@ -378,9 +381,11 @@ class BaseReportState:
         return pdf
 
     def _get_csv_data(self) -> bytes:
+        from flask import current_app
+
         url = self._get_url(result_format=ChartDataResultFormat.CSV)
         _, username = get_executor(
-            executors=app.config["ALERT_REPORTS_EXECUTORS"],
+            executors=current_app.config["ALERT_REPORTS_EXECUTORS"],
             model=self._report_schedule,
         )
         user = security_manager.find_user(username)
@@ -407,9 +412,11 @@ class BaseReportState:
         """
         Return data as a Pandas dataframe, to embed in notifications as a table.
         """
+        from flask import current_app
+
         url = self._get_url(result_format=ChartDataResultFormat.JSON)
         _, username = get_executor(
-            executors=app.config["ALERT_REPORTS_EXECUTORS"],
+            executors=current_app.config["ALERT_REPORTS_EXECUTORS"],
             model=self._report_schedule,
         )
         user = security_manager.find_user(username)
@@ -570,7 +577,9 @@ class BaseReportState:
             notification = create_notification(recipient, notification_content)
             try:
                 try:
-                    if app.config["ALERT_REPORTS_NOTIFICATION_DRY_RUN"]:
+                    from flask import current_app
+
+                    if current_app.config["ALERT_REPORTS_NOTIFICATION_DRY_RUN"]:
                         logger.info(
                             "Would send notification for alert %s, to %s. "
                             "ALERT_REPORTS_NOTIFICATION_DRY_RUN is enabled, "
@@ -880,8 +889,10 @@ class AsyncExecuteReportScheduleCommand(BaseCommand):
             self.validate()
             if not self._model:
                 raise ReportScheduleExecuteUnexpectedError()
+            from flask import current_app
+
             _, username = get_executor(
-                executors=app.config["ALERT_REPORTS_EXECUTORS"],
+                executors=current_app.config["ALERT_REPORTS_EXECUTORS"],
                 model=self._model,
             )
             user = security_manager.find_user(username)
